# contracts/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NoteRequest": {
        "properties": {
          "body": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "NoteRequest",
        "type": "object"
      },
      "NoteResponse": {
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "body",
          "author",
          "created_at"
        ],
        "title": "NoteResponse",
        "type": "object"
      },
...
      "ReviewRequest": {
        "properties": {
          "blind_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blind Id"
          },
          "confidence": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Confidence",
            "type": "integer"
          },
          "decision": {
            "enum": [
              "worth_deeper_analysis",
              "not_relevant",
              "uncertain"
            ],
            "title": "Decision",
            "type": "string"
          },
          "double_review_group": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Double Review Group"
          },
          "evaluation_split": {
            "anyOf": [
              {
                "enum": [
                  "development",
                  "validation",
                  "final"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Split"
          },
          "field_visit_performed": {
            "default": false,
            "title": "Field Visit Performed",
            "type": "boolean"
          },
          "protocol_arm": {
            "anyOf": [
              {
                "enum": [
                  "top_score",
                  "baseline",
                  "random"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol Arm"
          },
          "protocol_version": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol Version"
          },
          "rejection_reasons": {
            "items": {
              "enum": [
                "land_false_positive",
                "adverse_planning",
                "no_access",
                "risk_too_high",
                "estimate_too_optimistic",
                "works_too_large",
                "already_known",
                "outside_strategy",
                "other"
              ],
              "type": "string"
            },
            "maxItems": 9,
            "title": "Rejection Reasons",
            "type": "array"
          },
          "segment_code": {
            "anyOf": [
              {
                "enum": [
                  "metropolitan",
                  "medium_city",
                  "periurban",
                  "coastal",
                  "rural"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Code"
          },
          "strategy_code": {
            "anyOf": [
              {
                "enum": [
                  "division_extension",
                  "renovation_resale"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Code"
          }
        },
        "required": [
          "decision",
          "confidence"
        ],
        "title": "ReviewRequest",
        "type": "object"
      },
      "ReviewResponse": {
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "blind_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blind Id"
          },
          "confidence": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Confidence",
            "type": "integer"
          },
          "decision": {
            "enum": [
              "worth_deeper_analysis",
              "not_relevant",
              "uncertain"
            ],
            "title": "Decision",
            "type": "string"
          },
          "double_review_group": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Double Review Group"
          },
          "evaluation_split": {
            "anyOf": [
              {
                "enum": [
                  "development",
                  "validation",
                  "final"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Split"
          },
          "field_visit_performed": {
            "default": false,
            "title": "Field Visit Performed",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "protocol_arm": {
            "anyOf": [
              {
                "enum": [
                  "top_score",
                  "baseline",
                  "random"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol Arm"
          },
          "protocol_version": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol Version"
          },
          "rejection_reasons": {
            "items": {
              "enum": [
                "land_false_positive",
                "adverse_planning",
                "no_access",
                "risk_too_high",
                "estimate_too_optimistic",
                "works_too_large",
                "already_known",
                "outside_strategy",
                "other"
              ],
              "type": "string"
            },
            "maxItems": 9,
            "title": "Rejection Reasons",
            "type": "array"
          },
          "reviewed_at": {
            "title": "Reviewed At",
            "type": "string"
          },
          "reviewer_pseudonym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Pseudonym"
          },
          "segment_code": {
            "anyOf": [
              {
                "enum": [
                  "metropolitan",
                  "medium_city",
                  "periurban",
                  "coastal",
                  "rural"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Code"
          },
          "strategy_code": {
            "anyOf": [
              {
                "enum": [
                  "division_extension",
                  "renovation_resale"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Code"
          }
        },
        "required": [
          "decision",
          "confidence",
          "id",
          "author",
          "reviewed_at"
        ],
        "title": "ReviewResponse",
        "type": "object"
      },
...
      "ScenarioRequest": {
        "properties": {
          "fees_eur": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "default": "0",
            "title": "Fees Eur"
          },
          "finance_cost_eur": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "default": "0",
            "title": "Finance Cost Eur"
          },
          "holding_cost_eur": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "default": "0",
            "title": "Holding Cost Eur"
          },
          "purchase_price_eur": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Purchase Price Eur"
          },
          "resale_price_eur": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Resale Price Eur"
          },
          "works_cost_eur": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Works Cost Eur"
          }
        },
        "required": [
          "purchase_price_eur",
          "works_cost_eur",
          "resale_price_eur"
        ],
        "title": "ScenarioRequest",
        "type": "object"
      },
      "ScenarioResponse": {
        "properties": {
          "assumptions": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Assumptions",
            "type": "object"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "results": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Results",
            "type": "object"
          }
        },
        "required": [
          "id",
          "assumptions",
          "results",
          "created_at"
        ],
        "title": "ScenarioResponse",
        "type": "object"
      },
...
      "StatusRequest": {
        "properties": {
          "favorite": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favorite"
          },
          "rejection_comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Comment"
          },
          "rejection_reasons": {
            "items": {
              "enum": [
                "land_false_positive",
                "adverse_planning",
                "no_access",
                "risk_too_high",
                "estimate_too_optimistic",
                "works_too_large",
                "already_known",
                "outside_strategy",
                "other"
              ],
              "type": "string"
            },
            "maxItems": 9,
            "title": "Rejection Reasons",
            "type": "array"
          },
          "status": {
            "enum": [
              "new",
              "to_analyze",
              "retained",
              "contact_to_prepare",
              "contacted",
              "visit",
              "offer",
              "acquired",
              "lost",
              "rejected",
              "ignored"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Immo Opportunities API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/data-quality": {
      "get": {
        "operationId": "admin_data_quality_api_v1_admin_data_quality_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Admin Data Quality Api V1 Admin Data Quality Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Data Quality",
        "tags": [
          "connected-mvp"
        ]
      }
    },
    "/api/v1/admin/import-runs": {
      "get": {
        "operationId": "admin_import_runs_api_v1_admin_import_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 250,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Admin Import Runs Api V1 Admin Import Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Import Runs",
        "tags": [
          "connected-mvp"
        ]
      }
    },
...
    "/api/v1/opportunities/{opportunity_id}/notes": {
      "post": {
        "operationId": "create_note_api_v1_opportunities__opportunity_id__notes_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Note",
        "tags": [
          "connected-mvp"
        ]
      }
    },
    "/api/v1/opportunities/{opportunity_id}/reviews": {
      "post": {
        "operationId": "create_review_api_v1_opportunities__opportunity_id__reviews_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review",
        "tags": [
          "connected-mvp"
        ]
      }
    },
    "/api/v1/opportunities/{opportunity_id}/scenarios": {
      "post": {
        "operationId": "create_scenario_api_v1_opportunities__opportunity_id__scenarios_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scenario",
        "tags": [
          "connected-mvp"
        ]
      }
    },
...
    "/api/v1/opportunities/{opportunity_id}/status": {
      "patch": {
        "operationId": "update_opportunity_status_api_v1_opportunities__opportunity_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Opportunity Status",
        "tags": [
          "connected-mvp"
        ]
      }
    },
    "/api/v1/opportunities/{opportunity_id}/workspace": {
      "get": {
        "operationId": "opportunity_workspace_api_v1_opportunities__opportunity_id__workspace_get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Workspace",
        "tags": [
          "connected-mvp"
        ]
      }
    },
...
    "/api/v1/reviews/{review_id}/outcomes": {
      "post": {
        "operationId": "create_review_outcome_api_v1_reviews__review_id__outcomes_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewOutcomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review Outcome",
        "tags": [
          "connected-mvp"
        ]
      }
    },
    "/api/v1/saved-searches": {
      "post": {
        "operationId": "create_saved_search_api_v1_saved_searches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Saved Search",
        "tags": [
          "connected-mvp"
        ]
      }
    },
...
  }
}